Added all apis

+++ src/main/java/com/example/e_learning/repository/InstructorApplicationRepository.java
@@ -3,6 +3,6 @@ package com.example.e_learning.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.e_learning.entity.InstructorApplication;
 
-public interface InstructorApplicationRepository extends JpaRepository<InstructorApplication, Integer> {
+public interface InstructorApplicationRepository extends JpaRepository<InstructorApplication, Long> {
 	
 }
